requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
notebook>=6.5.0
openpyxl>=3.1.0
pyarrow>=12.0.0
google-cloud-secret-manager>=2.23.0